style(fjord-protocol): redundant lines and trailing comma

+++ packages/fjord-protocol/test/Wire.spec.ts
@@ -209,9 +209,7 @@ describe('0x05 - subscribe to torrent specific events', () => {
         uint8(ServerMessageType.TorrentEvent),
         hexString(hash),
         uint8(TorrentEvent.TorrentPiece),
-        uint32(1),
-        uint32(2),
-        uint32(1337),
+        uint32(1, 2, 1337),
       );
     });
   });
@@ -236,12 +234,8 @@ describe('0x05 - subscribe to torrent specific events', () => {
           uint8(ServerMessageType.TorrentEvent),
           hexString(hash),
           uint8(TorrentEvent.TorrentUpdate),
-          uint8(2),
-          uint8(25),
-          uint32(42),
-          uint32(50),
-          uint32(3),
-          uint32(4),
+          uint8(2, 25),
+          uint32(42, 50, 3, 4),
         ),
       );
     });
